Migliorata gestione del salto, input e animazioni

+++ GameEntity.h
@@ -83,7 +83,7 @@ protected:
     Hitbox hitbox;
     float movementSpeed;
     bool facingRight; //boolean value used to determine if the character is facing right or left
-    int movingCounter;//counter used to update sprite in order to generate walking/moving animation
+    float movingCounter;//counter used to update sprite in order to generate walking/moving animation
 };
 
 #endif //GAMEPROJECT_GAMEENTITY_H
